# Conflicts:
#	ProjectFiles/PupChaser/Assets/SaveScenes/Level1.unity

+++ ProjectFiles/PupChaser/Assets/Scripts/GoblinScripts/GoblinController.cs
@@ -134,7 +134,7 @@ public class GoblinController : MonoBehaviour {
 
 	public void killGoblin()
 	{
-		int i = Random.Range(1, 100);
+		float i = Random.Range(1, 100);
 		
 		if(i > 25)
 		{
